+++ Bruno/test.bru
@@ -4,8 +4,20 @@ meta {
   seq: 2
 }
 
-get {
-  url: http://localhost:3000
-  body: none
+post {
+  url: http://localhost:3000/api/users
+  body: json
   auth: none
+}
+
+headers {
+  content-type: application/json
+}
+
+body:json {
+    {
+          "username": "newuser",
+          "email": "newuser@example.com",
+          "password": "password123"
+      }
 }
