+++ app.js
@@ -2,10 +2,12 @@ const express = require('express');
 const app = express();
 const connection = require('./connection');
 require('dotenv').config();
-const nameRoute = require('./controller/nameRoute');
+const songRoute = require('./controller/songRoute');
+const userRoute = require('./controller/userRoute'); 
 
 app.use(express.json());
-app.use('/api', nameRoute);
+app.use('/api', songRoute);
+app.use('/api', userRoute); 
 
 app.get('/', (req, res) => {
     const dbStatus = mongoose.connection.readyState === 1 ? 'Connected' : 'Not Connected';
